third_party/blink/web_tests/platform/generic/netinfo/connection-types-expected.txt

Tests that all connection types of NetInfo are supported.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS connection.type is types[count][0]
PASS connection.downlinkMax is types[count][1]
PASS successfullyParsed is true

TEST COMPLETE
